+++ src/commands/configure.integration.test.ts
@@ -3,7 +3,7 @@ import {
   HttpScheme,
   SecurityType,
   SuperJson,
-} from '@superfaceai/sdk';
+} from '@superfaceai/one-sdk';
 import { join as joinPath } from 'path';
 import { mocked } from 'ts-jest/utils';