+++ src/commands/create.test.ts
@@ -1,5 +1,5 @@
 import { CLIError } from '@oclif/errors';
-import { SuperJson } from '@superfaceai/sdk';
+import { SuperJson } from '@superfaceai/one-sdk';
 import inquirer from 'inquirer';
 import { mocked } from 'ts-jest/utils';