+++ src/hooks/init.ts
@@ -1,7 +1,7 @@
 import { Hook } from '@oclif/config';
+import { VERSION as SDK_VERSION } from '@superfaceai/one-sdk';
 import { VERSION as PARSER_VERSION } from '@superfaceai/parser';
-import { VERSION as SDK_VERSION } from '@superfaceai/sdk';
 
 export const hook: Hook<'init'> = async function (_options) {
-  this.config.userAgent += ` (with @superfaceai/sdk/${SDK_VERSION}, @superfaceai/parser/${PARSER_VERSION})`;
+  this.config.userAgent += ` (with @superfaceai/one-sdk@${SDK_VERSION}, @superfaceai/parser@${PARSER_VERSION})`;
 };